move get_global_variables to variable_detector
