+++ src/detectmatelibrary/utils/persistency/component_interfaces.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel, ConfigDict, Field
 
 import os
 
-from detectmatelibrary.common._config._formats import _EventInstance, EventsConfig
+from detectmatelibrary.common._config._formats import EventsConfig
 from detectmatelibrary.schemas import ParserSchema
 from detectmatelibrary.tools.logging import logger
 
@@ -143,27 +143,6 @@ def get_configured_variables(
             if name in input_["logFormatVariables"]:
                 result[name] = input_["logFormatVariables"][name]
 
-    return result
-
-
-def get_global_variables(
-        input_: ParserSchema,
-        global_instances: Dict[str, _EventInstance],
-) -> Dict[str, Any]:
-    """Extract header variables from event-ID-independent instances.
-
-    Args:
-        input_: Parser schema containing logFormatVariables
-        global_instances: Dict of instance_name -> _EventInstance configs
-
-    Returns:
-        Dict mapping variable names to their values from the input
-    """
-    result: Dict[str, Any] = {}
-    for instance in global_instances.values():
-        for name in instance.header_variables:
-            if name in input_["logFormatVariables"]:
-                result[name] = input_["logFormatVariables"][name]
     return result
